requirements.txt
google-cloud-bigquery==0.25.0
pylint==1.7.2
termcolor==1.1.0
Flask=0.12.2
numpy==1.13.1
